pyproject.toml


[project]
name = "weather"
version = "0.1.0"
description = "I want to do some cool stuff with weather data"
authors = [
    {name = "Matt Livingston", email =  "mwl5399@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "requests >= 2.31.0",
    "pandas >= 2.1.4",
    "boto3 >= 1.34.153",
    "xarray>=2025.1.0",
    "grib2io>=2.3.0",
    "httpx>=0.28.1",
    "cdsapi>=0.7.5",
    "onepassword-sdk>=0.1.6",
    "asyncio>=3.4.3",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.5",
]
